Refactor loop indices and type handling for enhanced safety and consistency
- Replaced `int` with `size_t` in loops across `CDlgGrepReplace.cpp`, `CShareData.cpp`, and `CDlgFileTree.cpp`, improving compatibility with larger datasets and ensuring safer memory management.
- Continued to document type usage distinctions in AGENTS.md, providing clear guidelines for new code and reinforcing overall type consistency across the codebase.

+++ sakura_core/env/CShareData.cpp
@@ -82,7 +82,7 @@ CShareData::~CShareData()
 		CloseHandle( m_hFileMap );
 	}
 	if( m_pvTypeSettings ){
-		for( int i = 0; i < (int)m_pvTypeSettings->size(); i++ ){
+		for( size_t i = 0; i < m_pvTypeSettings->size(); i++ ){
 			delete (*m_pvTypeSettings)[i];
 			(*m_pvTypeSettings)[i] = nullptr;
 		}
